Split utils from `QuestionDataSource`
